Back out "Don't start server daemon threads by default"
Summary:
D52903768 caused problems because several cli commands (`index`, `write`, `finish`) do depend on the server threads and will hang if they are not started.

Moreover, several client APIs like `sendBatch` or `finishDB` depend on these background threads, which means it's not enough with enabling server mode for the cli commands that need it.

Reviewed By: simonmar

Differential Revision: D52954901

fbshipit-source-id: 5f44ba2b381771e4fdb292949d68ee3ffabc786f