relax the upper limit for call hierarchy
Summary:
Hitting the current limit 100 callers and annoyed that there isn't a way to override.

I think it is preferable to allow a higher  limit and just timeout if the cost is too high (there is already a 2s timeout installed).

Bump the default to 200 and remove the clamp from the explicit override

Reviewed By: simonmar

Differential Revision: D53084708

fbshipit-source-id: d9fa69ed727edf3f9a51edc427361b91376754b5