Store the directory of DBs in XDB
Summary:
Currently, getting the directory of glean DBs uses symlinks in Manifold to store which DBs are available. This is flaky and slow.

This diff uses XDB to store the list of available DBs.

I've opted to keep the manifold listing to be able to catch any bugs in my implementation. We can delete that and completely transition once we have enough confidence in this implementation.

Reviewed By: simonmar

Differential Revision: D52781810

fbshipit-source-id: d9ffa6243f0e6105483df9071e2e977e4a36fef7