Fix memory usage of numReferences
Summary:
1. `countApprox` didn't work, so remove it. It wasn't actually
running in constant space because I hadn't set a page size.  When I
did set a page size, the count became wrong because of duplicates. So
we really do have to track the actual results(*)
2. Add `searchWith` to search using an arbitrary `QueryResult` type
for transforming and combining results.
3. Use `searchWith` to make an efficient(ish) `numReferences`

(*) An alternative here would be to rely on using a local backend and
get Glean to eliminate all duplicates by having no limit on the
results. This doesn't work when using a server because the server will
always set a page size though, so I didn't pursue this.

Reviewed By: pepeiborra

Differential Revision: D53090512

fbshipit-source-id: 41922e1ad3fdfc85fd09e31dff848350a01ef7f8