implements Hack Require constraints
Summary:
This diffs modifies glass.thrift and glass to resurface the Hack "require *" constraints to users.

These constraints can be seen as a type of relation between entities and as such fit in the "SearchRelated" and "searchRelatedNeighborhood" services.

This adds 3 Glean queries to "searchRelatedNeighborhood" for Hack entities. They could be rewritten into a single query if needed. Incidentally, the [performance](https://fburl.com/scuba/glean_glass_server/28l2trxi) of SearchRelatedNeighborhood is pretty bad: P90 is 2.50s.

Reviewed By: simonmar

Differential Revision: D53088460

fbshipit-source-id: d966b898c329898b20e6cbe1c1acd2ee696cbfe5