runQuery: fetch the next page in parallel with decoding the results
Summary:
When running a large query with multiple pages, we wait a long time
decoding the current page before we send the request to the server for
the next page. By overlapping these we can save a lot of time
especially with bulk fetches.

Saves over 30% (3-4 minutes) in the inital phase of
`derive-function-calls` where we're fetching all the declaraitons.

Reviewed By: pepeiborra

Differential Revision: D53268246

fbshipit-source-id: 04b8a03f2e36f3e2691eed9fbae18f445dcfb365