Perform commit in parallel with rename
Summary:
This was fairly straightforward and seems to provide a nice speedup at
least for Python data: 24% faster writing a pile of JSON files. I
didn't see any difference for C++ but it's possible the gains would
only be visible for larger DBs.

One possible downside is increased memory usage because we can no
longer eagerly release the FactSet after writing, although we'll have
to test in prod to see what effect it really has.

Reviewed By: pepeiborra

Differential Revision: D52733037

fbshipit-source-id: 6518201cbda0a107e5f5ae583ede8f5e85f45e84