CxxDeclarationTargets: fetch FileXRefs in parallel
Summary:
Rather than fetching all the FileXRefs and XRefTargets up front, which
is slow and needs lots of memory, fetch them on demand for each file
as we process it. This has the effect of parallelising the fetching
when using multiple worker threads.

We already had a function to fetch the FileXRefs (used by
`derive-target-uses` and introduced when I was optimising that deriver
D51465090), in this diff I optimised it further:

1. Remove use of Haxl and replace with an explicit `PredMap`. Haxl was
slow for this use case.
2. Stream the FileXRefs rather than fetching them all. Some files have
a lot of FileXRefs and this saves a lot of memory.

This doesn't make a huge difference to performance, but it reduces memory
usage and enables more parallelism. Bigger wins coming later in the stack.

Reviewed By: pepeiborra

Differential Revision: D53228488

fbshipit-source-id: 6d730c5c85f68dbd89e4eb9b3d83227ab83c8051