Update C# schema with build-system data
Summary:
The motivation for the new predicates is to enable C# files to be opened in VSCode. This means that we must be able to look up which project (or solution) a file belongs to.

Not all projects however actually exist in the repo; for Unity, the projects are generated on the fly by Unity (or in the VSCode extension via the FAST generator).

The schema therefore models projects from two sources:
- MSBuildProjectSource - an MSBuild project (.csproj) checked into source-control
- UnityProjectSource - all the necessary information required to re-generate the required Unity project(s)

Stored predicates are included to re-index the unity-package<->project, project<->source, and solution<->project relations for search. Having both directions be indexed is important as we also want to be able to enumerate projects in a solution, etc. in a future "solution explorer" extension

Reviewed By: simonmar

Differential Revision: D53472616

fbshipit-source-id: a4b19a07b6ba129b7f70afd376694e520ec58542