Add some overloading to .|
Summary:
We often want to use `.|` in a statement context, but it's not
convenient for that because it only works on expressions. We have the
alternative `or_` but that's also inconvenient because you can't nest
it.

I've added some overloading so that `.|` can work in both an
expression and a statement context, which subsumes (and therefore
deprecates `or_`).

Reviewed By: josefs

Differential Revision: D53666766

fbshipit-source-id: 3423d99feb89636146a33b16fa7c14d075fb23c3