reinstate flaky integration test
Summary:
This test was disabled by simonmar due to timeout T177255162. Looking at the history, it was indeed running close to the 600s limit.

This test was slow because during its execution, it built from scratch ( isolation-dir...) glean and glass-snapshot.  This was useless since it's already builds them as dependencies of the test target.

This diff makes the test much faster by simply using the dependencies as resources, and also moves the test in glean/glass/facebook, along with glass-snapshot which has been moved there.

Reviewed By: pepeiborra

Differential Revision: D53685036

fbshipit-source-id: 2e75c869d9ef98f58e5863ce6d60fd8b3d2ce935