optimize generated code
Summary:
Improve OCaml generated code from Angle. The effect of this diff in the generated code can  be  seen in the next diff D53894807.

The current version is functionally correct, but make the hack indexer  ooms, which is why I didn't land D51967791.

Most likely, this is caused by using the OCaml std lib, rather than the more efficient "Core". I also got rid of the "filter_map" uses.

The code generated with the change is diff is now fully equivalent to the one in production and fixes the ooming issues.

Reviewed By: malanka

Differential Revision: D53895408

fbshipit-source-id: 3367de42c71e1979523ccbbd50fce6274dc3187d