Syncing Thrift changes from D54023581
Summary:
This diff only contains .thrift file changes, synced from {D54023581}.

I will stack a diff on top of this to remove `rust_crate_name` from TARGETS files. As of this diff, the Thrift files will contain `namespace rust` and the TARGETS files will contain `rust_crate_name`, but that is fine as long as the values are equal.

Reviewed By: zertosh

Differential Revision: D54044684

fbshipit-source-id: b84a247b84456363b584ac2f3f27141587dc5568