Refactor handler wrappers
Summary:
All this stuff is getting super complicated so I wanted to try to
clean it up.

* New module `Glean.Glass.Handler.Utils` for all the handler support stuff
* `selectRevision` is hidden
* `withStrictErrorHandling` is hidden
* `getGleanRepos`, `getSpecificGleanRepos` are hidden
* `withLog`, `withLogDB` are hidden
* We no longer pass around a `TVar GleanDBInfo`, instead we snapshot
  the `GleanDBInfo` in one place and pass it where it's needed. This
  means we always see a consistent view of DBs, probably fixes some
  minor race conditions. As a result, `lookupLatestDBs` is now pure, so I renamed it to `chooseGleanDBs`
* Lots of simplification and cleanup

I think this is a lot better but it's not as good as I
wanted yet: `withGleanDBs` should really go away, but I can't do that
without changing the semantics of `searchSymbol` which seems a bit
scary.

Reviewed By: pepeiborra

Differential Revision: D53813499

fbshipit-source-id: ee3fbfa1f6fa42ff0114e6820fc3f15e5b0412f8