Prioritise DB over snapshot when we have an exact revision match
Summary:
The DB is likely to have more accurate results, so use that if we have it.

This does potentially mean slowing down some requests since we wait
for both the snapshot and the Glean results before deciding which one
to use, so we should keep an eye on this in prod.

Reviewed By: pepeiborra

Differential Revision: D54200109

fbshipit-source-id: 77bdf71d7ac88537aea145f0841a0629c52eb19f