Remove identical duplicate comment spans
Summary:
If indexing produces duplicate entities with slightly varying structure (e.g. as occurs in C++ sometimes, with tiny variations in folly indexing resulting in copies of the entitity) we want to avoid generating the concatentation of their doc spans.

De-duplicate identical file/spans in the entities docs.

Underlying issues:
- too many copies of folly
- sometimes we concatenate, sometimes we uniq
- even though we have the fbcode folly::future entity, we pick up the _comments_ for the xplat one. DeclFamily issue?

Reviewed By: simonmar

Differential Revision: D54234753

fbshipit-source-id: e5b8470e40a39b5374badaa3544e39f392b456e6