Suport a language Id flag to the encoder
Summary:
Some SCIP indexers don't set the LanguageId of files. This is the case for scip-typescript

Glean needs to know the language id at encoding time, so that scip symbols get the right tag.

set it here and it will get propagated into the json before we write to glean.

Reviewed By: alichtman

Differential Revision: D54524549

fbshipit-source-id: f694a6a8217beb9cd608b2d76395a27fbf9c6141