codemarkup predicates for xlang navigations
Summary:
We map generated entities (in the "referring" db, e.g. fbsource) to `idlEntity` in the "foreign" db (e.g fbsource.fbthrift, using new predicate `codemarkup.EntityIdl`

The `idlEntity` is produced from the referring db. Consequently, it may miss information about the idlEntity (typically, won't include the location).

We can't join predicates from two different dbs in a derived predicate. If the location isn't provided, glass needs to resolve the location by querying the idl DB.

There are three increasing levels of precision an idlEntity can offer
1. idl entity file. Glass can return it to the user as a xref pointing to the start of file
2. idl entity fact. Glass can generate a symbol id,  providing using access to symbol api doc (and indirectly go-to-definition). It can get the location with an extra glean call to the idl db
3. a location.  Enough for glass to provide all xref info without additional Glean calls, relying only derived predicates.

For Thrift, we have 1 and 2. I didn't see if 3 is possible. This design leaves this option open but doesn't commit to it.

Notes: C++ vs other languages. Even though the initial POC is for C++ which has it's own logic, we want to be sure the proposed approach works for other languages.

Reviewed By: pepeiborra

Differential Revision: D53063132

fbshipit-source-id: 9c3fdadcebf0454564494121eaeecdedad0d2e79