add option to include xlang_xrefs
Summary:
Add a request option so it's not enabled by default. This'll provide more control on deployment.  Make it a feature flag initially, will see if we want to make it default or add a specific request option instead. In any case, it only impact the the documentSymbolList and documentSymbolIndex methods.

1. Clients don't expect  extra xrefs (for a given source span). Note there are already such cases. In hack, dynamic dispatch, in C++ decl/def. Codehub/Phab picks an arbitrary xref.

2. Latency can be impacted too since there's an extra glean call, but it's executed in parallel to other calls, and most of the time will return 0 results.

Reviewed By: pepeiborra

Differential Revision: D55001342

fbshipit-source-id: 0a8d59f49766d57945e5ace6d1626029d7c7bc3e