generate thrift xrefs for C++
Summary:
DocumentListSymbol and DocumentSymbolIndex methods use the predicates defined in D53063132 in order to compute the idl xrefs. This is limited to the C++ path and other methods aren't affected.

It generates the new xrefs by performing an extra query to `CxxFileEntityIdl`. The new xrefs are returned along with the other "regular" xrefs, with a specific attribute to tell them apart.

I factorized the XRef type which is shared between Handler and Cxx, and updated it so it can represents either type of xref (idl and regular).

Idl locations aren't resolved.
- They can be either provided as annotation in the referring db
- or resolved by glass in follow up D55044706
- In this diff, we use a dummy value, so they'll point to the first line of the thrift file

## Discussion

 Idl xrefs are resolved by a call to `codemarkup.cxx.CxxFileEntityIdl` (with a FileXRefs as key, just like for the other C++ queries). This is done concurrently to other methods so impact on latency should be limited, and most of the time it returns empty results.

I think it's possible to avoid this extra call by extending the existing C++ queries (fixed/variable xrefs) with an extra `codemarkup.EntityIdl` clause. This'd also allow us to get rid of `codemarkup.cxx.CxxFileEntityIdl`.

As a first step, I think the approach in this diff is simpler as it requires less changes. I'll keep the optimization as a follow up.

Reviewed By: pepeiborra

Differential Revision: D53479706

fbshipit-source-id: 6cada965b49084ff2ac71e1aa9c6e8b442a1fe74