use cell for annotation
Summary:
The thrift path that appears in the thrift generated code annotations is relative to the repo "cell", which can be fbcode.  For instance,  `data_preproc/perf_test/DummyTrainerIf.thrift` instead of `fbcode/data_preproc/perf_test/DummyTrainerIf.thrift`

But in the cxx.CxxToThrift predicate, we want a path from fbsource.

This diff
- adds the cell (of the target of) the source file being indexed to the visitor state <- at least this is what I hope it is "source.target"
- update the thrift files' path

Reviewed By: josefs

Differential Revision: D55473403

fbshipit-source-id: fbbfad63f446a45bdc47552c14420c85582c9815