move facebook specific out of open-source code
Summary:
This addresses simonmar's comment in D55473403. It solves the OSS issue but  this `fbsource` condition is still wrong.

I couldn't see a simple way to fix this.  I need to understand the interplay of all options  (--root, --cwd_subdir, --target_subdir, --path_prefix) and also how the target passed to the indexer are constructed. For instance, the target in the tests cases are of the form "fbsource//glean/..."...

It may be simpler to generate the thrift paths relative to repo-root directly in the thrift generation.

Reviewed By: simonmar

Differential Revision: D55658413

fbshipit-source-id: 91fe4eac3142a754fdb5c8f99104cf0727d816ce