telemetry (#440)
Summary:
Pull Request resolved: https://github.com/facebookincubator/Glean/pull/440

Traces are written to Scuba glean_glass_server_events and visualised with Chrome (Perfetto)

{F1477527085}
https://fburl.com/scuba/glean_glass_server_events/gkhqrmoo

Right now only the following things are instrumented:

* The top level Thrift Handler
* documentSymbolsList/Index

RFC:

* would this be useful to debug Glass performance on individual requests?
* any concerns about running it in production?

One limitation is tracing inside Haxl computations. Would need a bit more work to instrument the Haxl datasource. Once that's done, it would theoretically be possible to carry the instrumentation through to the Glean server and have whole-system traces. But that's probably overkill.

I'll clean up the code and test the OSS build when/if there is consensus to land it

Reviewed By: josefs

Differential Revision: D55698252

fbshipit-source-id: ccd7c41345fcb2a54e5c3de6ed68912797516e6d