SCS cache
Summary:
A really basic cache. We're expecting tends of thousands of revisions,
so cache size is not really an issue. I went with a bounded cache that
has approximate LRU by keeping two semi-spaces, kind of like a 2-space
GC. It's suboptimal in space and time but scores highly in simplicity.

Reviewed By: nhawkes

Differential Revision: D55800094

fbshipit-source-id: d2655a6e5d3b5d886e159d200fe22024f4098af1