Retire the "attributes" database mechanism from Glass
Summary:
About 3 years ago when we were shipping codehub navigation for the first time, one of the blockers was showing gCPU data from Xenon on hover cards.
This was a feature in Diffusion that we had to match to allow CodeHub Browser to ship.

The problem was the xenon data only knew Hack "class::method" qnames. While in Glean we had entities. We had no way at the time to stack the www.perf.hack data over the Hack db, so instead Glass did a client-side "join" on document symbols.

This meant opening an additional attributes db, then doing a bulk fetch of any extra metadata by qname.
The safety was enforced by a type level mechanism that associates a key in the attribute db with an entity in the base db.

I think now days we would add this extra metadata at indexing time (ideally), where the Hack indexer would grab the daily xenon data as a final step in indexnig.
Alternativley, you'd stack the perf data over the base Hack db and join on entities.

We don't need this stuff anymore since the www.perf.hack db has gone away and no one cared.
We _should_ revisit how to store Xenon data properly.

Reviewed By: simonmar

Differential Revision: D56339981

fbshipit-source-id: da55c060cc3112cf3784f7b34ebc390f8590dc66