Update schema with member-access x-ref
Summary:
This diff adds support to the C# schema for capturing 'member access' expressions, e.g.

 {F1494280477}

This allows us to wire up a new x-ref for the variable/parameter being accessed (in this example, `someLocalVariable`).

In this iteration we will index member access expressions that are part of method-invocation expressions (other member-access x-ref sites will be added in future as well as other alternatives for `MemberInvocationExpression`).

Reviewed By: malanka

Differential Revision: D56465646

fbshipit-source-id: e0df1dae66212eae446d6b4e98d199a04018d1d0