No XRefs for method calls in Python
Summary:
Python unit tests of XRefs are limited to `FileEntityXRefLocations`.
This relation (and I believe all XRef relations in Python) omits
method references.  There are in fact Entities defined for methods,
and there's an opportunity to at least point to a generic method name
entity in the absence of disambiguating information.

This currently makes it impossible to find method calls in Python code
in the diff sketcher.

By the way, did you know there are three identical tests of the same
query in this directory, testing xrefs in the rather sparse all.py?
This is one of them.  I switched it to main.py which at least calls a method.

Reviewed By: simonmar

Differential Revision: D55526693

fbshipit-source-id: f05fb55f33385fc5400805bda35d59c81fe9d225