Produce Cxx.DeclarationTargets in the indexer
Summary:
The deriver for `cxx1.DeclarationTargets` used source ranges to infer
the set of references from a declaration, but this gives wrong results
in a number of ways:

1. The data was incorrect for definitions in macro expansions
2. The data was incorrect for records: references from an inline
method were included in the parent record's references, but not
out-of-line methods.
3. The data was incorrect when there were local definitions
inside a function body.

To fix these things we have to generate the facts from the AST and not
the source ranges, which is what the diff does.

Other benefits:

* Now we get `cxx1.DeclarationTargets` in the incremental DB too
* One less manual deriver, lots of ugly code deleted
* Faster indexing (perhaps?)

But doing this led to perf issues because we now have to de-duplicate
a lot of `cxx.DeclarationTargets` facts for header files. Initial
versions of this diff generated vast amounts of extra data (like 3x
more `cxx1.DeclarationTargets`). Mitigated by:

* Filter out references to `NamespaceDeclaration`. I don't think these
were useful anyway.
* Canonicalise references (point to definitions rather than
declaraitons if available). This reduces differences between the
facts, allowing more de-duplication.

We still get a lot of different versions of `cxx1.DeclarationTargets`
for some records in headers - just like xrefs - but overall we're
producing less data than before so I think this is better.

The test output differences are all due to either (a) references from
inline methods are no longer counted as references from the parent
record, and (b) no more references to/from namespaces.

Reviewed By: mpark

Differential Revision: D56349548

fbshipit-source-id: 32c10d72d26595312a10a2d2858922d9eb99ab99