Make FragmentSpread a predicate on graphql.angle
Summary:
High-level goal: support go-to-def from fragment spread to fragment in CodeHub through glass integration.

This diff makes FragmentSpread a predicate, and changed the corresponding indexer code to write it. That way FragmentSpread has the location info on it to power go-to-def.

Note: I did not go through the formal process of changing the version of the schema, because the GraphQL schema in glean is unused. Let me know if that's required though!

Reviewed By: pepeiborra, ginfung

Differential Revision: D57071967

fbshipit-source-id: ba45982352633fdcdad81c2a197a9e1a1003f302