Retire the bespoke lionhead.lionizer attribute parameter lookup schema
Summary:
This hasn't returned a result in > 90 days.
The last and only client is the lionhead harness generator, and that support was removed in the previous diff.
We can now remove this schema.

That should unblock some docmarkup.angle removal

lionead.lionizer itself is a on-going thing, just not this parameter lookup system.  fbsource/fbcode/security/lionhead/lionizer

Reviewed By: pepeiborra

Differential Revision: D57367774

fbshipit-source-id: 0902693445909cce086166862a9e3a3ddd38e03f