Add predicate mapping src.File to build config
Summary:
This is needed because different build configs might contain query/fragments of the same name in fbsource (example: https://fburl.com/code/0gknsuu8). For go-to-def, we need to to make sure we are jumping to the fragment from the same build config.

I decided to map `src.File` to build config because this is the largest unit that does't get split between configs, and we can map any type containing location info to a config using the new predicate.

Reviewed By: ginfung

Differential Revision: D57473420

fbshipit-source-id: 4112fac86a0a04478d9cdc9c135e5edd7e2f9393