Trace DB restore steps
Summary:
We emit events for the whole restore and some of the internal steps. If a step throws, the trace will show the exception details.

This is partly intended as an example of how we can add more telemetry to the Glean server to diagnose issues and have more visibility of what the server is doing. If nothing else, it's a more visual form of logging.

Reviewed By: simonmar

Differential Revision: D57627262

fbshipit-source-id: 3fe0d1fbee5d3b8d8b9a40b2ba8784d8db4fc994