Fetch the snapshot in parallel with Glean queries
Summary:
Fixes a recent regression.

Unfortunately we do end up possibly fetching the content hash for the
requested revision twice with `matching_revision`: once in the Glean
branch and once in the snapshot branch (in parallel). There's no easy
way to avoid this, because we don't know whether we need it until we
have found the Glean DB containing the file and/or found the
snapshot. Perhaps we could add some caching if this is a problem.

Reviewed By: pepeiborra

Differential Revision: D57622949

fbshipit-source-id: 613202c072e497aa3e8bd121f21e94b66de9334b