Refactor DB retention
Summary:
Extract a pure and testable component to calculate a set of retention
changes for the local server: which DBs to delete and restore.

We already moved in this direction by extracting `computeRetentionSet`
earlier, this diff moves more of the logic into the pure component and
isolates it in a separate module behind a nice clean API.

There was no need for the `StateT` stuff to be fixed by the pure
component, so I made it fully abstract over `Monad m`.

Reviewed By: josefs

Differential Revision: D57724154

fbshipit-source-id: 2d2f13e1de9db46509bd7fcf946a5d5a9f9f4e04