Add a few more languages to the language tag in glean.scip.encode
Summary:
Some SCIP indexers don't set the LanguageId
https://github.com/sourcegraph/scip/blob/main/scip.proto#L756

field in the scip output. This makes it hard to guess what language the symbol is in. Since SCIP is a multi-language format we actually need to know at least at the db level. For Java, we need it at the symbol level as Kotlin and Java code exist in the same repo.

This adds a few more cases to the scip encoder, used prior to writing blobs to the db to encode scip to Angle.

Reviewed By: phlalx

Differential Revision: D58120224

fbshipit-source-id: f07d102a856469a9298f34a61c3952ccd9e30bcb