Check in the generated cpp schema.
Summary:
Prior to this diff, we were generating the C++ schema on-the-fly via `glean/schema/gen/Glean/Schema/Gen/Thrift.hs`.

We bundle this generation into the `glean/schema/sync` workflow. The main motivation for the change is due to not being able to run a Haskell binary on Windows. Rather than trying to port the generation Haskell code, we check it in instead here.

Reviewed By: malanka, simonmar

Differential Revision: D55511581