Check in the generated cpp schema. (#452)
Summary:
Pull Request resolved: https://github.com/facebookincubator/Glean/pull/452

Prior to this diff, we were generating the C++ schema on-the-fly via `glean/schema/gen/Glean/Schema/Gen/Thrift.hs`.

We bundle this generation into the `glean/schema/sync` workflow. The main motivation for the change is due to not being able to run a Haskell binary on Windows. Rather than trying to port the generation Haskell code, we check it in instead here.

Reviewed By: malanka, simonmar

Differential Revision: D55511581

fbshipit-source-id: 5f9a1fe296440c54976e35c6ecd28688e6d7cc86