Use Meta in the Site class
Summary:
The `Site` class uses the type `Map String String` to store meta properties. But this is a kind of lie. It contains a single entry `"meta"` with the meta properties being stuffed into a single JSON object which is then serialized to string. This is fairly horrible to work with.

This diff uses the `Meta` type in the interface instead. That turns out to clean up a whole bunch of code and error checking.

This is a breaking change for implementations of the `Site` class, but they should be easier to implement now, in my experience.

Reviewed By: simonmar

Differential Revision: D58014303

fbshipit-source-id: d5d23209918a194d16ce507e811c5aa4fb33c553