Deshim AutoTimer in glean
Summary:
The following rules were deshimmed:
```
//folly/experimental:auto_timer -> //folly/logging:auto_timer
```

The following headers were deshimmed:
```
folly/experimental/AutoTimer.h -> folly/logging/AutoTimer.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_deatx.glean

Reviewed By: yfeldblum

Differential Revision: D58148285

fbshipit-source-id: 87e0eb9c3304b43efa1b682c92d486cc3130ec4a