Java entities now found via SCIP
Summary:
We 'de-register' the native Java and Kotlin schema entities in favour of the SCIP-embedded Java and Kotlin symbols.

Glass only knows how to find language entiies under one branch of code.Entity. This might be a problem at some point but for now it just means explicitly changing the way we search from one to the other entity tags.

A general solution would be to search both .

Reviewed By: simonmar

Differential Revision: D58508972

fbshipit-source-id: 53e0a7d46799cf0402b82990cc8c8e68060c7945