Retire codemarkup.FileEntityXRefSpans
Summary:
This was a early attempt at target-less xrefs. However, it was lingering in the IsIndexed predicate, where its actively misleading.

SCIP languages (Rust, TypeScript and Java) do not have spans, only ranges. Also true for some cxx1 / pp types. As a result files only containing xrefs for those languages would appear to not be indexed with xrefs.

Just delete and fix the use to refer to FileEntityXrefRangeSpans

Reviewed By: simonmar

Differential Revision: D58501467

fbshipit-source-id: 93717358186edc3a1692afa5202dbc6cc9a4d831