refactor SearchEntity
Summary:
`symbolSearch`  resolves symbol ids into entities + location.

It is used in two contexts
1. find-references
2. describe

The location is needed only for 2.

In the context xlang references, we may have referenced entities in a db for which we don't have a location. In that case, the `symbolSearch` API won't be able to return an entity.

The goal of this refactoring is to not  constraint `symbolSearch` to return locations.  This is needed in D58248829.

The location isn't part of a `SearchResult` and has to be made explicit.

Types such as `(SearchResult Cxx.Entity)` become `(SearchResult (ResultLocation Cxx.Entity))`

Reviewed By: simonmar

Differential Revision: D58199759

fbshipit-source-id: 83ae465b1226d1f2b754a7383744b664b8674a5b