Reduce O(n^3) query to O(n^2)
Summary:
The `scopeQ` query was cubic in the number of nested namespaces, which
leads to OOMs.

This has been known for a long time (T149694011) but it seems that
traffic patterns changed and we've been receiving more of the bad
requests lately, causing a lot of OOMs and service degradation
(S426006).

This reduces the query to only quadratic. I did try to make it linear,
but it turns out to be quite tricky (it's an interesting problem, I
will probably keep trying). But this should alleviate most of the
problems.

Reviewed By: malanka

Differential Revision: D58666945

fbshipit-source-id: 127274e9488a884768b76456c050e0c58fc8dfe9