search fields
Summary:
This diffs adds the ability to search/decode symbol ids for fields. This builds on top of the current logic which is flawed and should be rewritten.

Thrift symbol ids are of the form

path/service_name       # service
path/type_name            # exception, enum, struct
path/file                         # file (included file)
path/name                     # constant
path/service/function   # function
path/enumname/enum # enum
path/typename/field     # union, struct, exception field

This is ambiguous, arity alone isnt enough to infer the proper case.

The decoder is best effort and returns the first match based on the order it runs queries. It's both inefficient (require several queries) and incorrect as we may be unable to decode some symbol ids.

To fix that, we need to add an extra tag for disambiguation in the symbol id, and deterministic decoding.

Reviewed By: simonmar

Differential Revision: D59060840

fbshipit-source-id: c9801a875052be8ea138bab389373af52e008636