To prepare for the Thrift compiler to check for reserved identifier names in the Thrift IDL, add the annotation to the generated thrift files that need that allows them to use the reserved identifier name
Summary:
The change includes the code changes to generate those files.

The reason we need to go through this step is because of the sync to www. Sync to www needs the files to be in master before they are synced to www. Without the annotation, the build will fail in www and will not allow the compiler change to land. Land the files with the annotation ahead of the compiler change to prevent the build failure.

Reviewed By: praihan

Differential Revision: D59134460

fbshipit-source-id: 84d4c0103f9a30422203c8b4cb730fbe7602e50d