no need to patch hack symbol description
Summary:
A glass symbol full description has a field to store the symbol it overrides https://fburl.com/code/u3y3psn5

This field is set using a SearchRelated query  which, in the Hack case, resolves to `codemarkup.hack.HackExtendsParentEntity`, and eventually to `hack.Overrides`.

Until recently, the `hack.Overrides` predicate was generated from the `override` annotations, leading to missing override data. To circumvent this, glass was patching the missing annotation from the class hierarchy information it computes.

This is no longer needed since D57452291.

Reviewed By: donsbot

Differential Revision: D59700249

fbshipit-source-id: 30d6c3cba4b4b256e8bcc5cd939d84b69f640261