Cleanup how enum constants work
Summary:
The problem I want to solve is when looking at an upper-case identifier like `Class_`, the Angle compiler has to decide whether `Class_` refers to an enum constant or a variable. It needs to know the type of the enclosing context, which isn't always available. So we have confusing situations where it's OK to write `Class_ = Variable` but not `Variable = Class_`.

I first tried to solve this by adding new syntax. But that leads to a painful migration and we don't really want new syntax. So the new solution is

1. Soft-deprecate upper-case enum constants. It'll still work on a best-effort basis (which is what we do now) and we will try not to break it.
2. Fully support lower-case enum constants. These are now parsed into a different AST node (`Enum`) and don't rely on type-directed syntax in the compiler.

Maybe in the future we could add a warning. For now I'll update the docs (later diff).

Reviewed By: josefs

Differential Revision: D59469942

fbshipit-source-id: d5e7dae840a43118ab535e19775667ef1362c5d0