Don't use inheritedN parents at all in Hack case
Summary:
We don't need to return empty containers, as the UI doesn't render them (in fact, they're filtered out in the Hack layer). Specialize the Hack case specifically to just the InheritedEntities search.

EntShare seems to have quite a large number of such classes, so we save some time not computing describe() for them.

This is worth about 15% runtime for some of the larger ents.

Some of the inheritance graphs involve up to 1300 empty parents.

Reviewed By: phlalx

Differential Revision: D59740231

fbshipit-source-id: 3873e7afd96a3ee53c8a1bc5174572f213349ab7