Add debug flags
Summary:
Optional debug logging is something we don't do very well at the
moment in Glean. Using `vlog` is quite awkward because we don't have
the same facilities to turn it on/off per module in the same way that
C++ does. We could implement that, but I think a nicer way will be to
have debug flags accepted by the backend, this is a start at that.

Reviewed By: malanka

Differential Revision: D59757576

fbshipit-source-id: d3348838bbaebde7a7bb50f3e850573fb5cdaf48