Use an arm VM directly for CI (#461)
Summary:
The github provider arm64 runners don't have docker on them, so if we try to use a 'container' stanza to pick a particular linux version, we fail. We don't actually need to use a nested vm since the host is a ubuntu 22 image on arm64 anyway. So just run natively.

needs `sudo` passwordless for apt-get tho

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/461

Reviewed By: simonmar

Differential Revision: D59899791

Pulled By: donsbot

fbshipit-source-id: ee869ecac9c16da7e47b1dffb507316c7839c682