Type Inference
Summary:
I'll write a post about this in due course but to summarise what's going on here: the idea is to fix the quirks in Angle that currently require you to write things in a particular order to keep the type checker happy. Most often this comes up as the need to write `Y = X` rather than `X = Y`, or write statements in a particular order, or write a type signature.

There are three main techniques being used
1. Type variables and unification. This is fairly standard stuff found in type inference for functional languages, e.g. https://en.wikipedia.org/wiki/Hindley%E2%80%93Milner_type_system
2. Extensible record and union types. This is using a form of [refinement types](https://en.wikipedia.org/wiki/Refinement_type), e.g. `Has { x : nat } Y` means "Y is a type that has at least the field x of type nat".
3. A small constraint solver for promoting key types to predicate types. (see the `
Note [Promotion]` in the comments in Typecheck.hs).

A nice consequence of this is we can handle primops properly, several of which have polymorphic types, e.g. `prim.concat : forall a . [a] -> [a] -> [a]`.

Reviewed By: josefs

Differential Revision: D59686685

fbshipit-source-id: a92e3deabb3ec8cec24878e1fbd2d8a5e70b8f2e